🚚 Move dev files
