+++ genet/_dev/__init__.py
@@ -1,4 +1,4 @@
-from genet.predict_dev.functional_dev import *
+from genet._dev.functional_dev import *
 from silence_tensorflow import silence_tensorflow
 import warnings
